% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grspwr.R
\name{allelesToGenotypes}
\alias{allelesToGenotypes}
\title{Combine single alleles into genotypes}
\usage{
allelesToGenotypes(alleles)
}
\description{
pairs diploid-alleles assorted according hardy-weinberg equibrilium principles iff alleles are in random order
}
