% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grspwr.R
\name{grspwr}
\alias{grspwr}
\title{GRS power calculation}
\usage{
grspwr(snps, n, alpha = 0.05, max.iter = 1000, popsize = max(n)^2)
}
\arguments{
\item{n}{The number of individuals to include in the GRS}

\item{max.iter}{The number of iterations to use for the power calculation. This affects the precision of the calculated power. Should be at least 100.}

\item{popsize}{The size of the population sample from which is sampled. Should be larger than N. Affects accuracy of power estimate.}

\item{snps:}{A dataframe with the following columns:
\itemize{
\item SNP: Unique name of SNP
\item WEIGHT: a weight associated with the _minor_ allele of SNP. Weights should be normalized   relative to a standard normal distribution (mean 0, variance 1).
\item EAF: Effect allele frequency
\item Estimated R-squared correlation of imputation with actual genotype.For imputed genotypes this will be [0-1] and for directly genotyped SNPs it should be 1.
}}
}
\description{
Calculates the power to detect of a GRS to detect a signal at given significance level
using a GRS  weights in a certain sample size and with given imputatation quality.
Note: It assumes that the weights used are accurate estimates of SNP effects.
}
\examples{
exampleGRS <- data.frame(SNP = c("rs1", "rs2", "rs3"),
                              WEIGHT = c(0.01, 0.01, 0.2),
                              EAF = c(0.42, 0.42, 0.42),
                              INFO= c(0.9, 0.9, 0.2))
grspwr(exampleGRS,n=400, alpha = 0.05)
}
