% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grspwr.R
\name{optimizedPermutation}
\alias{optimizedPermutation}
\title{Permute genotypes subject to opmization}
\usage{
optimizedPermutation(genotypes, fn)
}
\arguments{
\item{genotypes}{A vector of genotypes}

\item{fn}{A function which takes a vector of genotypes as first argument and return a score.}
}
\description{
Generates a permutation of genotypes that minimizes fn.
}
