% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grspwr.R
\name{randomSwap}
\alias{randomSwap}
\title{Swap two genotypes}
\usage{
randomSwap(genotypes)
}
\arguments{
\item{genotypes}{A vector of genotypes}
}
\description{
Swaps a random pair of indices in a vector of genotypes
}
