% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grspwr.R
\name{sampleDosages}
\alias{sampleDosages}
\title{Sample dosages from genotypes}
\usage{
sampleDosages(genotypes, info, epsilon = 0.01)
}
\arguments{
\item{genotypes}{a vector of numeric genotypes (0,1,2)}

\item{info}{Desired \eqn{R^2} correlation between dosages and genotypes}

\item{epsilon}{A halting parameter: When the total absolute difference
between desired and obtained \eqn{R^2} reaches epsilon the dosages are returned.}
}
\description{
Creates random dosages from genotypes using fast binary search for r-squared.
It will result in  approximately correct r-squared.
}
