% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grspwr.R
\name{sampleGenotypes}
\alias{sampleGenotypes}
\title{Generate genotypes with a given effect on phetypes}
\usage{
sampleGenotypes(phenotypes, beta, n, maf)
}
\arguments{
\item{phenotypes}{a vector of phenotypes (assumed to be normally distributed)}

\item{beta}{desired effect size}

\item{n}{sample size}

\item{maf}{minor (effect) allele frequency}
}
\description{
Make a vector of genotypes with a certain frequency and such that the
per-allele effect one the given phenotypes is as specified.
}
