% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grspwr.R
\name{scaleDosages}
\alias{scaleDosages}
\title{Scale a numeric vector to values between 0 and 2.}
\usage{
scaleDosages(unscaled)
}
\arguments{
\item{unscaled}{The unscaled input}
}
\description{
Scale a numeric vector to values between 0 and 2.
}
