% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grspwr.R
\name{sampleGenotypes}
\alias{sampleGenotypes}
\title{Generate a genotype vector of length n with a given minor allele frequency (maf)}
\usage{
sampleGenotypes(n, maf)
}
\description{
Generate a genotype vector of length n with a given minor allele frequency (maf)
}
